done with reorganization, made pages folder with home and about pages
