+++ package.json
@@ -9,6 +9,7 @@
     "bootstrap": "^5.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
   },
