+++ src/App.js
@@ -1,0 +1,17 @@
+import React, { Fragment } from 'react';
+
+import Home from "./pages/Home";
+
+import './App.css';
+
+function App() {
+
+  return (
+    <Fragment>
+      <Home />
+    </Fragment>
+  );
+
+}
+
+export default App;
