+++ src/components/Header/Header.css
@@ -5,8 +5,6 @@
     margin: 0 auto;
 }
 
-
-
 button {
     width: 100px;
     height: 40px;
