+++ src/pages/About.js
@@ -1,0 +1,12 @@
+import React, { Fragment } from "react";
+
+import "./About.css";
+
+const About = () => {
+    
+    return(
+        <Fragment>
+            
+        </Fragment>
+    );
+}
